+++ cordformation/src/main/kotlin/net/corda/plugins/Dockerform.kt
@@ -99,6 +99,7 @@ open class Dockerform @Inject constructor(objects: ObjectFactory) : Baseform(obj
                             "$nodeBuildDir/additional-node-infos:/opt/corda/additional-node-infos",
                             "$nodeBuildDir/drivers:/opt/corda/drivers"
                     ),
+                    "environment" to listOf("ACCEPT_LICENSE=\${ACCEPT_LICENSE}"),
                     "ports" to listOf(it.rpcPort.get(), it.config.getInt("sshd.port")),
                     "image" to dockerImage.get()
             )