+++ changelog.md
@@ -4,6 +4,7 @@
 
 ### Version 5.0.10
 
+* `cordformation`: Add option to accept license agreements for Corda enterprise Docker images 
 * `cordformation`: Add properties `runSchemaMigration` and `allowHibernateToManageAppSchema` in order to run schema 
 migration/set-up as part of a cordform deployment (only supported in Corda 4.6 and later)
 * `cordformation`: Made Dockerform's dockerImage property mandatory to avoid using hardcoded default Docker images